This merge brings in the following changes from upstream:

* 9c87466c3960: This patch changes the OpenMP runtime to call `abort`
  instead of `exit` on failures.  A side effect is that OpenACC
  Profiling Interface shutdown callbacks are no longer dispatched
  after a failure from offloading.  Similar to our upstream
  9b4826d18b5f for OpenMP's test suite, this merge updates Clacc's
  test suite to expect the new failure mode.

* 932316660179: This patch adjusts the interface of `DeviceTy`.  This
  merge updates Clacc's OMPT device implementation accordingly.

* At this merge, Clacc CI is seeing frequent racy failures when
  compiling OpenACC programs for ppc64le, both without any offloading
  and with offloading to nvptx64.  However, it's challenging to
  reproduce the failures in a shell.  We have not identified the
  breaking commit, but we believe it's an upstream change for ppc64le,
  and we hope future upstream changes will fix this once we merge
  them.

+++ openmp/libomptarget/src/rtl.h
@@ -58,7 +58,7 @@ struct RTLInfoTy {
       uint64_t, __tgt_async_info *
       OMPT_SUPPORT_IF(, const ompt_plugin_api_t *ompt_api));
   typedef int64_t(init_requires_ty)(int64_t);
-  typedef int64_t(synchronize_ty)(int64_t, __tgt_async_info *);
+  typedef int64_t(synchronize_ty)(int32_t, __tgt_async_info *);
 
   int32_t Idx = -1;             // RTL index, index is the number of devices
                                 // of other RTLs that were registered before,
